[otbn, isa] Define vectorized bignumber instructions
This defines the encoding and description of SIMD instructions
which work on unsigned elements placed in the WDRs.

The ISA extension is designed to support vector element sizes
of 16, 32, 64 and 128 bits. However, this semester thesis
will focus on only 16 and 32 bit elements.

The added instructions are:
- addv / addvm
- subv / subvm
- mulv(.l) / mulvm.(l)
- trn1 / trn2
- shv

These instructions are grouped into
the opcode b1011011 named custom4.

Encoding is also prepared for future instructions like:
addvc, subvc, addvi, subvi

Signed-off-by: Pascal Etterli <petterli@student.ethz.ch>