[otbn, syn] Benderize OpenTitan Repo for OTBN synthesis
This adds all files for OTBN synthesis to the bender manifest.
The synthesis is based upon the Earlgrey ASIC but during synthesis/elaboration
we select OTBN as target. Thus we need all packages but only the OTBN sv files
and its direct dependencies.

FYI: Bender is a dependency management tool used by ETHZ.

Signed-off-by: Pascal Etterli <petterli@student.ethz.ch>